refactor(gh): convert gh.rs to a directory module, extract gh/stars
src/gh.rs (1435 lines, 67 top-level pub fns, no internal module
boundaries) is being split into per-resource submodules over a
5-PR series (issue #301). This PR does the one-time src/gh.rs ->
src/gh/mod.rs conversion and extracts the first (smallest) resource:
gh/stars.rs, holding the batched-GraphQL stargazer-count functions
and their 3 unit tests.

escape_graphql_string stays in gh/mod.rs as shared kernel (also used
by a not-yet-extracted fork-flags query, moving in a later PR).
gh/mod.rs re-exports stars' public items by explicit name (matching
tui/mod.rs's existing pub-use convention, not a glob), declared near
the bottom of the file alongside the eventual mod tests block (same
placement convention tui/mod.rs uses) -- so none of the ~30 existing
crate::gh::X call sites need to change.

The remaining 4 planned submodules (gists, comments, forks,
revisions) still live inline in gh/mod.rs; later PRs in this series
extract them one at a time.

No signature or behavior change.

Part of #301
